Split off PsiFile and VirtualFile util
