+++ src/nl/hannahsten/texifyidea/util/files/VirtualFile.kt
@@ -1,0 +1,85 @@
+package nl.hannahsten.texifyidea.util.files
+
+import com.intellij.openapi.project.Project
+import com.intellij.openapi.vfs.LocalFileSystem
+import com.intellij.openapi.vfs.VirtualFile
+import com.intellij.psi.PsiFile
+import com.intellij.psi.PsiManager
+import nl.hannahsten.texifyidea.util.appendExtension
+import java.io.File
+
+/**
+ * Looks up the PsiFile that corresponds to the Virtual File.
+ */
+fun VirtualFile.psiFile(project: Project): PsiFile? {
+    if (!this.isValid) return null
+    return PsiManager.getInstance(project).findFile(this)
+}
+
+fun VirtualFile.findVirtualFileByAbsoluteOrRelativePath(filePath: String): VirtualFile? {
+    val isAbsolute = File(filePath).isAbsolute
+    return if (!isAbsolute) {
+        findFileByRelativePath(filePath)
+    }
+    else {
+        LocalFileSystem.getInstance().findFileByPath(filePath)
+    }
+}
+
+/**
+ * Looks for a certain file, relative to this directory or if the given path is absolute use that directly.
+ *
+ * First looks if the file including extensions exists, when it doesn't it tries to append all
+ * possible extensions until it finds a good one.
+ *
+ * @param filePath
+ *         The name of the file relative to the directory, or an absolute path.
+ * @param extensions
+ *         Set of all supported extensions to look for.
+ * @return The matching file, or `null` when the file couldn't be found.
+ */
+fun VirtualFile.findFile(filePath: String, extensions: Set<String> = emptySet()): VirtualFile? {
+    val isAbsolute = File(filePath).isAbsolute
+    var file = if (!isAbsolute) {
+        findFileByRelativePath(filePath)
+    }
+    else {
+        LocalFileSystem.getInstance().findFileByPath(filePath)
+    }
+    if (file != null && !file.isDirectory) return file
+
+    extensions.forEach { extension ->
+        val lookFor = filePath.appendExtension(extension)
+        file = if (!isAbsolute) {
+            findFileByRelativePath(lookFor)
+        }
+        else {
+            LocalFileSystem.getInstance().findFileByPath(lookFor)
+        }
+
+        if (file != null && !file!!.isDirectory) return file
+    }
+
+    return null
+}
+
+/**
+ * Recursively finds all files in a directory (thus, also the files in sub-directories etc.)
+ */
+fun VirtualFile.allChildFiles(): Set<VirtualFile> {
+    val set = HashSet<VirtualFile>()
+    allChildFiles(set)
+    return set
+}
+
+/**
+ * Recursive implementation of [allChildFiles].
+ */
+private fun VirtualFile.allChildFiles(files: MutableSet<VirtualFile>) {
+    if (isDirectory) {
+        children.forEach {
+            it.allChildFiles(files)
+        }
+    }
+    else files.add(this)
+}
